+++ core/systemctrl/src/loadercore.c
@@ -297,9 +297,6 @@ SceModule2* patchLoaderCore(void)
             case 0x30894000:    _sw(0x3C090000, addr);               break;        // Allow Syscalls
             case 0x00E8282B:    _sh(0x1000, addr + 6);               break;        // Remove POPS Check
             case 0x01A3302B:    _sw(NOP, addr+4);                    break;        // Remove Invalid PRX Type (0x80020148) Check
-            //case 0x5040FF98:    // Remove beqzl
-		    //case 0x5040FF54:    // are these really needed?
-		    //case 0x5040FF2E:    _sw(NOP, addr); _sw(NOP, addr+4);    break;
             }
         }
     }